Use a string rather than a []byte to register i18n tables
This makes the code way faster to parse by go fmt and also saves
a lot of space in the source code (it's around 50% smaller). It might
also reduce startup time, since the compiler probably copies the data
when instantiating the []byte (because it's modificable), but not
when instantiating the string (it's not modificable, so it can just
use a pointer to the DATA section).