Add support for template nodes in the compiler and HTML escaping
Currently we're using reflect and unsafe to get a pointer to the
escaping functions, since they're unescaped. This is only
a temporary solution. Then plan is to copy the relevant functions
from the html/template package to allow the compiler to run in
environments which don't allow unsafe and also to let us optimize
how the escaping functions are used (currently they ignore the
output type of the previous function, but this information could
be used to gain performance.

Right now, the compiler is around 2x faster the the executor from
the go std library and produces exactly the same output.