Always use compiled templates
This way we avoid having to do so much mangling with the trees,
and initial tests show that in a full web page, the VM is about
10x faster than the interpreter (and we still can optimize the VM
a bit more). Furthermore, this also allows us to decide if we
should escape the template based on its content type. This way
Gondola templates can be used for both text and html.