Remove JMPE instruction and enhance iteration interface
Rather than returning -1 when iteration ends, return a boolean
which indicates if iteration has ended. Perform the check and
conditional jump in the NEXT instruction, so we avoid stacking
the values and false positives about iteration end (previously,
a map with int keys and containing -1 would cause the iteration
to stop).