Initial support for the debugging framework
The debugging framework is conditionally compiled with the debug
build tag. This adds very minimal overhead (a call to
runtime.deferreturn for functions which didn't use defer) when
not enabled. Currently, it measures the time spent in cache, orm
and templates, but users are able to add their own measurements.