Clean up watching code and fix issues on Linux
As it turns out, the fsnotify is quite a leaky abstraction, since
the behavior of watcher.Watch() varies by OS. Watching only GOROOT
and GOPATH does not work on Linux, since the watch is per-dir, not
including subdirs. Also, clean up the code which filters events,
moving the extensions which are considered source code to a slice.