Correctly call Orm.Initialize on GAE
App.Orm() will now return non-err when using GCSQL. This way we
can use the same code to call Orm.Initialize() from App.Prepare().
Note that this wouldn't work with the datastore, since it requires
an incoming request to be able to talk to the DS. However, since
the DS does not require table creation, this solution works for now.