Remove SqlQuery() from Orm for now.
It's ugly and difficult to understand, we'll come up with a better
replacement. Also, avoid add a Connection() method to the Driver
interface, so it's possible to get the connection to the underlying
database for every driver. This also lets us type assert on
driver.Connection() rather than on the driver itself when creating
the ORM to obtain a *sql.DB if the driver is SQL based.