A few incompatible changes to arithmetic functions in templates
We now have add, sub and mul (note it lost the t at the end).
There are 'naked' variants, which return an interface which might
be either a float64 or an int, depending on if the result has a
fractional part.
For each function, there's a i and a f version, which always
returns an integer or a float64, respectivelly.
Finally, the functions int and float have been added.