Change the API for gnd.la/form
Make gnd.la/form the primary API.
Renderer packges, like gnd.la/bootstrap register themselves as the
default renderer on import.
This allows libraries to just use gnd.la/form and lets the final
user to choose the renderer package which matches the frontend
framework used for the site.

TODO: Do this for the pager too.