Cleanup in gnd.la/app
Remove app.Cache and app.Orm, use just cache.Cache and orm.Orm. Those
types introduced too much confusion for little gain. If the user
mistakenly closes the orm or the cache of the app, it should be easy
to find. This allows us to clean up the initialization code
significantly.