Change paginator.Paginator to accept the total number of items and the items per page
This ends up saving code on the caller's side, since most of the
time the caller will have a total count and a fixed number of
items per page, rather than the number of pages. In those cases,
the caller can just pass 1 as the second argument and the paginator
will work fine anyway.